backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
python-dotenv
feedparser
openai
